+++ shared/GR_Dataset.lua
@@ -22,6 +22,8 @@ function ds:WhatsNew() -- What's new in the current version
     |CFFFFFF00v3.0.38 Notes|r
         - Added option to keep addon running and ignore certain ways to close it.
         - esMX (Spanish Mexico) localization added.
+        - Fixed issue with whispers not being sent when player joins guild.
+        - Fixed issue with compact mode not removing anti-spam/black list before adding to invite list.
     |CFFFFFF00v3.0.36 Notes|r  -- Joined forces with FGI to bring recruiting to the next level.
         - Changed the 6 months to 180 days.
         (NOTE: If you were using the 6 months, you will need to update your settings.)